+++ app/src/main/java/vn/mbm/phimp/me/leafpic/activities/SettingsActivity.java
@@ -1,5 +1,7 @@
 package vn.mbm.phimp.me.leafpic.activities;
 
+import android.app.Fragment;
+import android.app.FragmentManager;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -35,6 +37,8 @@ import vn.mbm.phimp.me.leafpic.util.ThemeHelper;
 
 import uz.shift.colorpicker.LineColorPicker;
 import uz.shift.colorpicker.OnColorChangedListener;
+import vn.mbm.phimp.me.opencamera.CamUtil.CameraActivity;
+import vn.mbm.phimp.me.opencamera.CamUtil.MyPreferenceFragment;
 
 import static vn.mbm.phimp.me.leafpic.util.ThemeHelper.AMOLED_THEME;
 import static vn.mbm.phimp.me.leafpic.util.ThemeHelper.DARK_THEME;
@@ -103,6 +107,17 @@ public class SettingsActivity extends ThemedActivity {
                 else
                     askPasswordDialog();
 
+            }
+        });
+        /*** CAMERA ***/
+        findViewById(R.id.ll_camera).setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View v) {
+
+                MyPreferenceFragment fragment = new MyPreferenceFragment();
+                fragment.setArguments(CameraActivity.bundle);
+                getFragmentManager().beginTransaction().add(R.id.pref_container, fragment, "PREFERENCE_FRAGMENT").addToBackStack(null).commitAllowingStateLoss();
+                findViewById(R.id.settingAct_scrollView).setVisibility(View.GONE);
             }
         });
 
@@ -810,4 +825,24 @@ public class SettingsActivity extends ThemedActivity {
         ((TextView) findViewById(R.id.map_provider_item_sub)).setTextColor(color);
         ((TextView) findViewById(R.id.media_viewer_swipe_direction_sub)).setTextColor(color);
     }
+
+    private void setLayout(){
+
+    }
+
+    @Override
+    public void onBackPressed() {
+        FragmentManager fm = getFragmentManager();
+        if (fm.getBackStackEntryCount() > 0) {
+            fm.popBackStack();
+            if (fm.getBackStackEntryCount()==0)
+                findViewById(R.id.ll_camera).setVisibility(View.GONE);
+                findViewById(R.id.settingAct_scrollView).setVisibility(View.VISIBLE);
+
+        }
+        else {
+            super.onBackPressed();
+        }
+    }
+
 }
