+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/HDRProcessorException.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 /** Exception for HDRProcessor class.
  */
