+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/ImageSaver.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 import android.annotation.SuppressLint;
 import android.annotation.TargetApi;
