+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/MyDebug.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 /** Global constant to control logging, should always be set to false in
  *  released versions.
