+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/MyTileService.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 import android.content.Intent;
 import android.os.Build;
@@ -9,9 +9,9 @@ import android.util.Log;
 /** Provides service for quick settings tile.
  */
 @RequiresApi(api = Build.VERSION_CODES.N)
-public class MyTileServiceVideo extends TileService {
-    private static final String TAG = "MyTileServiceVideo";
-    public static final String TILE_ID = "vn.mbm.phimp.me.opencamera.TILE_VIDEO";
+public class MyTileService extends TileService {
+    private static final String TAG = "MyTileService";
+    public static final String TILE_ID = "vn.mbm.phimp.me.opencamera.TILE_CAMERA";
 
     @Override
     public void onDestroy() {
