+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/PreferenceKeys.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 /** Stores all of the string keys used for SharedPreferences.
  */
