+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/SaveLocationHistory.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 import android.content.SharedPreferences;
 import android.preference.PreferenceManager;
