+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/TextFormatter.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 import android.content.Context;
 import android.location.Location;
