+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CamUtil/ToastBoxer.java
@@ -1,4 +1,4 @@
-package vn.mbm.phimp.me.opencamera;
+package vn.mbm.phimp.me.opencamera.CamUtil;
 
 import android.widget.Toast;
