+++ app/src/main/java/vn/mbm/phimp/me/opencamera/CameraController/CameraController.java
@@ -9,7 +9,7 @@ import android.media.MediaRecorder;
 import android.util.Log;
 import android.view.SurfaceHolder;
 
-import vn.mbm.phimp.me.opencamera.MyDebug;
+import vn.mbm.phimp.me.opencamera.CamUtil.MyDebug;
 
 import java.util.List;
