+++ app/src/main/java/vn/mbm/phimp/me/opencamera/Preview/VideoQualityHandler.java
@@ -4,7 +4,7 @@ import android.media.CamcorderProfile;
 import android.util.Log;
 
 import vn.mbm.phimp.me.opencamera.CameraController.CameraController;
-import vn.mbm.phimp.me.opencamera.MyDebug;
+import vn.mbm.phimp.me.opencamera.CamUtil.MyDebug;
 
 import java.io.Serializable;
 import java.util.ArrayList;
