+++ app/src/main/res/values/strings.xml
@@ -535,6 +535,8 @@
 
     <string name="type_password">Insert password</string>
     <string name="security_sub">Set up a password to protect your media.</string>
+    <string name="camera_sub">Manage all setting about camera
+    </string>
     <string name="remember_password_message">Note: password changed, if you forget it you will have to clear application data.</string>
 
     <string name="apply_password_on">Use password for</string>
